pyproject.toml
[project]
name = "fastapiexample"
version = "0.1.0"
description = "Simple example crud project for entities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "confz>=2.0.1",
    "fastapi>=0.115.5",
    "loguru>=0.7.2",
    "psycopg2>=2.9.10",
    "pydantic>=2.10.2",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.32.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "pytest-xdist>=3.6.1",
    "ruff>=0.8.1",
]

[tool.mypy]
python_version = 3.12
strict = true
disable_error_code="import-untyped"
follow_imports="silent"

[tool.ruff]
line-length = 88
indent-width = 4
lint.select = ["F", "W", "B", "C4", "Q", "I"]
lint.ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"